Work on subtractorMOG
changed to constructor without params.  Works better but app closes
after 10 min or so.  No exception.  Also worked to get full resolution.
Works but with hack in camerabridgebase to fit image to canvas.
Otherwise preview is a cropped portion of 1920x1028.  Now testing with
basic detector to see if it does the same thing.